requirements.txt
streamlit
supabase
openai
pinecone-client
requests
python-dotenv
PyPDF2
pytest
pytest-mock